refactor directory structure
