--- server2/src/main/kotlin/com/junktion/api/v1/models/image/Image.kt
@@ -1,9 +1,0 @@
-package com.junktion.api.v1.models.image
-
-import java.time.LocalDateTime
-
-data class Image(
-		val tag: String,
-		val createdAt: LocalDateTime,
-		val updatedAt: LocalDateTime
-)
