--- server2/src/main/kotlin/com/junktion/api/v1/models/image/ImageFileStorage.kt
@@ -1,7 +1,0 @@
-package com.junktion.api.v1.models.image
-
-import java.io.File
-
-interface ImageFileStorage {
-	fun save(file: File, tag: String)
-}
