--- server2/src/main/kotlin/com/junktion/api/v1/models/image/ImageService.kt
@@ -1,8 +1,0 @@
-package com.junktion.api.v1.models.image
-
-import java.io.File
-
-interface ImageService {
-	fun upload(): String
-	fun upload(file: File, tag: String)
-}
