+++ server2/src/main/kotlin/junktion/v1/api/ImageFileStorage.kt
@@ -1,0 +1,7 @@
+package junktion.v1.api
+
+import java.io.File
+
+interface ImageFileStorage {
+	fun save(file: File, tag: String)
+}
