+++ server2/src/main/kotlin/junktion/v1/api/ImageRepository.kt
@@ -1,0 +1,7 @@
+package junktion.v1.api
+
+import junktion.v1.core.Image
+
+interface ImageRepository {
+	fun register(image: Image)
+}
