+++ server2/src/main/kotlin/junktion/v1/api/ImageService.kt
@@ -1,0 +1,8 @@
+package junktion.v1.api
+
+import java.io.File
+
+interface ImageService {
+	fun upload(): String
+	fun upload(file: File, tag: String)
+}
