+++ server2/src/main/kotlin/junktion/v1/core/Image.kt
@@ -1,0 +1,9 @@
+package junktion.v1.core
+
+import java.time.LocalDateTime
+
+data class Image(
+		val tag: String,
+		val createdAt: LocalDateTime,
+		val updatedAt: LocalDateTime
+)
